phil: added usleep(100) for ARG='3 310 103 103' and changed time int to long long

+++ Curriculum/08_philosophers/01_code/do_eat.c
@@ -6,7 +6,7 @@
 /*   By: akekesi <akekesi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/06 20:06:29 by akekesi           #+#    #+#             */
-/*   Updated: 2023/09/08 22:08:42 by akekesi          ###   ########.fr       */
+/*   Updated: 2023/09/09 03:10:48 by akekesi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,8 +41,8 @@ void	do_eat(t_phil **phil)
 
 void	get_fork_left(t_phil **phil)
 {
-	int				time_0;
-	int				time_1;
+	long long	time_0;
+	long long	time_1;
 
 	time_0 = get_time_current();
 	while (1)
@@ -64,8 +64,8 @@ void	get_fork_left(t_phil **phil)
 
 void	get_fork_right(t_phil **phil)
 {
-	int				time_0;
-	int				time_1;
+	long long	time_0;
+	long long	time_1;
 
 	time_0 = get_time_current();
 	while (1)
